django-server/requirements.txt
asgiref==3.8.1
boto3==1.34.138
botocore==1.34.138
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.13.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
